+++ src/mem/stack.rs
@@ -2,9 +2,14 @@ use std::alloc::Layout;
 use std::mem::MaybeUninit;
 use crate::mem::{Mem, MemBuilder};
 
-/// Fixed capacity on-stack memory.
+/// Fixed `SIZE` capacity on-stack memory.
 ///
-/// `SIZE` in bytes.
+/// Can contain at most `SIZE` bytes.
+/// Upon `build()` real capacity calculated based on `element_layout`.
+/// This involves `idiv` operation. You may want to consider [`StackN`]
+/// for intermediate storage instead.
+///
+/// [`StackN`]: super::StackN
 #[derive(Default, Clone)]
 pub struct Stack<const SIZE: usize>;
 impl<const SIZE: usize> MemBuilder for Stack<SIZE>{
@@ -12,16 +17,25 @@ impl<const SIZE: usize> MemBuilder for Stack<SIZE>{
 
     #[inline]
     fn build(&mut self, element_layout: Layout) -> StackMem<SIZE> {
+        let size =
+            if element_layout.size() == 0{
+                usize::MAX
+            } else{
+                SIZE / element_layout.size()
+            };
+
         StackMem{
             mem: MaybeUninit::uninit(),
-            element_layout
+            element_layout,
+            size
         }
     }
 }
 
 pub struct StackMem<const SIZE: usize>{
     mem: MaybeUninit<[u8; SIZE]>,
-    element_layout: Layout
+    element_layout: Layout,
+    size: usize
 }
 
 impl<const SIZE: usize> Mem for StackMem<SIZE>{
@@ -42,6 +56,6 @@ impl<const SIZE: usize> Mem for StackMem<SIZE>{
 
     #[inline]
     fn size(&self) -> usize {
-        SIZE
+        self.size
     }
 }